services/kb/requirements.txt
# Core Web Framework (Latest as of Jan 2025)
fastapi[standard]>=0.128.0
uvicorn[standard]>=0.40.0

# AI & LLM (Latest versions)
openai>=2.15.0
sentence-transformers>=5.2.0
torch>=2.9.1

# Vector Database (Latest Pinecone SDK v8)
pinecone>=8.0.0

# Environment & Utilities
python-dotenv>=1.2.1
python-multipart>=0.0.21

# Document Processing
pypdf2>=3.0.1

# Database
supabase>=2.27.2
